[iOS] Privacy setting UI changes for Safe Browsing
In the Privacy settings menu, the safe browsing row is currently in the
same section as Clear Browsing Data. To have UI as intended, we need to
separate them into different sections. Additionally the footer was
moved to the bottom of all the rows as that is where it should be
placed. There was also a period missing in a standard protection
sub page string so I also decided to put that change in here since it's
a related small nit.

UI Screenshot:
https://drive.google.com/file/d/1lfr-tCw-zy04DjkJu-NKn0be6X1ivQ17/view?usp=sharing&resourcekey=0-RioWIW4zCWGE2ORMa7xb8A

Bug: 1295870
Change-Id: I3012cfaa9fb07638b968104c7155820355e3f75c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3611067
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997166}
NOKEYCHECK=True
GitOrigin-RevId: 717f257e47ffb068d5eb3107e0a0bda218584d67