[ios] Pass SequencedTaskRunner to DownloadTaskImpl
In preparation for refactoring of how download destination file
is created (and of fixing bug 929112), change DownloadTaskImpl
to get a SequencedTaskRunner to use for all background jobs.

This allow all concurrent download to run their background tasks
on the same sequence, and more importantly, for DownloadTaskImpl
and its sub-class to use the same sequence for their background
tasks.

Bug: 929112
Change-Id: I7ae3873d2f274371b77403a71d82cb9159406b6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3607959
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997669}
NOKEYCHECK=True
GitOrigin-RevId: fd99b9e0fe25e1e7bb557ff77bca0571fcf3cf8f