Clear shopping item meta when bookmark URL changes
This patch adds a BookmarkModelObserver to the shopping service to
watch for URL changes on nodes that have shopping meta. In this case
the meta is cleared since we can't currently guarantee that the new
URL is also a shopping page.

The subscription will be removed the next time the user is on their
android device as a result of the preexisting validation logic.

This patch also adds the shopping service to both android and ios
service lists as it now has a function.

Bug: 1304895
Change-Id: Ib76451f50187292ec843d4ee1b97a2bb2ce679a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3573585
Reviewed-by: Wei-Yin Chen <wychen@chromium.org>
Auto-Submit: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Zhiyuan Cai <zhiyuancai@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#998376}
NOKEYCHECK=True
GitOrigin-RevId: 58a87bc0bfec22095f035df7250b35bad06ee282