HTTPS-Only Mode: Add UI setting on iOS
This CL tentatively adds an option under Privacy Settings to enable the
HTTPS-Only Mode on iOS. At the moment, the option isn't wired to the
actual feature, so it doesn't do anything. It will also be moved to the
new Security Settings when it's available.

The existing feature flag controls the display of this option, so it's
not immediately visible to users.

Bug: 1302509
Change-Id: If999c5b81a4b10377f2cafaf1f9cc6bae98a487e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3589164
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#999097}
NOKEYCHECK=True
GitOrigin-RevId: 4de8863a6a56b16f4f54028f2c316d4d3c16c057