[ios] Change DownloadTask file ownership
Remove the support for downloading to memory and instead always
download to a file. If the path given to `DownloadTask::Start()`
is empty, then the download will be saved to a temporary file
(removed when the task is destroyed).

Remove the `destination_hint` parameter from the `Start` method as
it is now ignored (as the download will always proceed to disk).

Move the file creation (and selection of the temporary location in
case of an empty path) to `DownloadTaskImpl::Start()` method. This
method will ensure that the directory structure exists, that the
path does not correspond a directory, and that no file exist at the
path, and that a file can be created.

Change `GetResponseData()` to be asynchronous and to read the data
from the downloaded file and convert the two tab helper that used
this method to support the asynchronous operation.

Remove most of the logic not related to performing the download out
of DownloadTaskImpl sub-classes that now only need to override two
methods (with a third one optional).

Adapt the sub-classes and the clients to the new API. One notable
change is that it is no longer possible to read the downloaded data
while the download is in progress (this has always been racy anyway
and was only used in tests).

Bug: 929112
Change-Id: I0bb6c5cd8c880b192830e6f981058206fb698264
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3610096
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#999312}
NOKEYCHECK=True
GitOrigin-RevId: cb6149ee37351a0bf03c2f3c30269326ebc145bf