HTTPS-Only Mode: Stop the timer before showing interstitials
When we upgrade an HTTP navigation to HTTPS, we start a timer that
checks if the upgraded navigation is taking too long. When the timer
expires we show the interstitial.

The code presently triggers an interstitial in a few places but it
doesn't properly stop the timer in all those places. This CL fixes
that and checks to see if the timer is still running in tests.

Bug: 1302509
Change-Id: I139c267c0d6faa839ca882bca9e8f1997bbb464a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3627857
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#999654}
NOKEYCHECK=True
GitOrigin-RevId: f6e31a732c148ff255687e09fd5b78ea658b3772