[iOS] Remove last suggestion line separator in SwiftUI omnibox popup
This CL disables the custom separator of the last row in a section in
the SwiftUI omnibox popup. A new bottom separator for the whole popup is
added as a replacement, similar to the bottom separator implemented in
`omnibox_popup_presenter.mm'. Thus, the latter is now disabled when the
SwiftUI version of the popup is the one presented.

Fixed: 1322386
Change-Id: If13b5d6a1be46864ea139bd520635299176731ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3629353
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Reviewed-by: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1000300}
NOKEYCHECK=True
GitOrigin-RevId: 234c9431c23d1b4adb637552bd6bd3f7c4b3f4df