[iOS] Omnibox popup: pedal debouncer shouldn't prevent closure.
The pedal debouncer accidentally introduced a bug when the popup is
closing.
To close the popup, its contents are set to an empty list of
suggestions. However, if the last result before closure had a pedal,
the debouncer makes it stick around for 0.3s, effectively preventing the
popup from closing.

Add a special case for this, and a test to make sure this can't degrade.

Bug: 1322915
Change-Id: Id375728955427f2d2cd7ee5fd57b815b989dd86b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3629235
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1000345}
NOKEYCHECK=True
GitOrigin-RevId: 1bbb2225cdf253f1bf04188b7f1e90060cd7415a