[iOS] Reduce height of some buttons
The goal is that FRE’s button, in default size, becomes 50px. It was
tested with font size at default value (i.e. middle of the slider in
iOS accessibility).

Result can be seen on
https://drive.google.com/file/d/1tIVY7IDCkoArmXpGXRqyreVSkqK5BJ5R/view?usp=sharing

Measuring on an emulator with point-perfect size, using both the
screenshot tool and the "print description" of the button in the view
hierarchy.

This also change the buttons using the same inset, that is,
"confirmation alert view", "passwords in other apps view", and
"sc_first_run_hero_screen_view".

Primary button reduced by 5 px (reducing inset). Also, in primary and
secondary buttons of FRE, font size is restricted to fit in the width
of the button, up to .7.

Bug: 1313164
Change-Id: If8203d26a61c469a7f42ea20738c4de0ff77a41b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3596762
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Louis Romero <lpromero@chromium.org>
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1000363}
NOKEYCHECK=True
GitOrigin-RevId: 73ac7e671981e9b7c1016e9c861fb650d6ec2aba