[iOS] Chrome iOS FRE small devices: bug fixes on SignInSyncVC
This CL fixes some layout overlapping issues when the user uses larger
text on iOS device. Before this change, the bottom anchor of the
identity picker and the top anchor of the enterprise caption are not
explicitly defined, causing the overlap.

This CL also updates the enterprise caption, which used to be of size
UIFontTextStyleFootnote. It is now updated to UIFontTextStyleCaption2.

Bug: N/A
Change-Id: I89106eba860bc972ec2ed4eb341d7d3de06e7e43
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3632074
Commit-Queue: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1000966}
NOKEYCHECK=True
GitOrigin-RevId: 99109fb02c914ab8c09403ecff56799636adfe75