[iOS] Chrome iOS FRE small devices: Only show all buttons when scrolled
PromoStyleViewController should be updated so that when
self.scrollToEndMandatory is set to true, secondary and tertiary buttons
are hidden until the user scrolls to the bottom.

In case that when secondary/tertiary buttons appears and hid the end of
the content, the "read more" tap handler will keep scrolling up to the
end to reveal it. A 0.05s delay is imposed to make sure the buttons are
displayed and the bounds of the scroll view are properly updated so the
bottom position of the scroll view content is accurate.

Fixed: 1323101
Change-Id: Ia15c99113651d4911c0210903b704a21b5077dff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3631563
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1001047}
NOKEYCHECK=True
GitOrigin-RevId: a51009dda17e8d4cb4fac22353a410d1ccb60c0e