Factor ActivityServiceCommands out of BrowserCommands
As part of separating the command protocols to make BrowserCommands
better-defined and easier to use with HandlerForProtocol(), this CL
removes the ActivityServiceCommands protocol from BrowserCommands.

The primary user of this protocol is ToolbarButtonActionsHandler, which
is updated to have separate per-protocol command handlers.

Since there is no longer a code path where the BVC is created and loads
its view without having a browser and browser state set, the conditional
call to addUIFunctionalityForBrowserAndBrowserState is now unconditional,
and some of the code in that method is hoisted into -viewDidLoad so the
ordering is correct for setting up command handling.

Similarly, the toolbar button factory is moved to be initialized after
the location bar, since it (now) requires OmniboxCommands to be usable.

Bug: 1045047
Change-Id: Ie3de78e1bc1f83dfaf14cd438c5b369a21aef206
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3613320
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1001075}
NOKEYCHECK=True
GitOrigin-RevId: 07ef5978001823ef69e69dedd3905d009329d078