[iOS] Update sublabel color in Treatment 2 of SwiftUI popup
This CL changes the SwiftUI omnibox popup row suggestion detail text
color in treatment two. Specifically, "entity" suggestions used to use
the same color for both the suggestion title and the detail text, namely
`kTextPrimaryColor`. If treatment two is enabled, it will now treat
entity suggestions like other suggestions and set the detail text color
to `kTextSecondaryColor`.

`PopupMatchRowView` is cleaned-up so as to reflect the fact that it is
not responsible for setting the suggestion text color anymore, since
attributed strings are now used. The only exception to this is when a
row is highlighted, in which case we set the text color to white.

Pedal suggestion wrappers, which also provide attributed strings, are
now responsible for setting the text color through the foreground color
attribute.

Fixed: 1323323
Change-Id: I8f248935ccd1c4f1f8345a4ef6e1134ebfb7b4de
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3637993
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1001518}
NOKEYCHECK=True
GitOrigin-RevId: 01e34f6e6d1a113153c61ffd751bf647f56a4b9a