[iOS] Chrome iOS FRE small device: iPad font size and margin adjustments
This CL does the following:

1. Fixes existing bug that the title label of PromoStyleViewController
on iPad becomes UIFontTextStyleLargeTitle instead of
UIFontTextStyleTitle1. The original issue is caused by the size class of
`self.traitCollection` is "compact" when the view controller is
presented on a form sheet; this CL updates the implementation so that
the size class is computed using its presentation controller instead of
itself, if available.

2. Reduce the left and right margins so that the content spans across
80% of the form sheet, instead of 65%.

3. Fix some button margin bugs introduced by crrev.com/c/3631563

Fixed: 1323172,1324162
Change-Id: I4c1bfcb83e4ba72ac19eeb6c90fdba2af58a78ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3635662
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Ginny Huang <ginnyhuang@chromium.org>
Auto-Submit: Ginny Huang <ginnyhuang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1001611}
NOKEYCHECK=True
GitOrigin-RevId: d5acbf3d567fa19dd5b005bda3fe7090d4a0cdce