HTTPS-Only Mode: Make HTTP allowlist per-profile instead of per-tab
HTTPS-Only Mode has an allowlist that stores the user decisions when
the user clicks through an HTTPS-Only Mode interstitial. If the user
navigates to the same domain later, this allowlist is checked to decide
whether to allow the navigation over HTTP or to upgrade it to HTTPS.

The allowlist is currently per-tab, meaning it doesn't carry over to
other tabs. To fix this, this CL introduces a new KeyedService called
HttpsUpgradeService which replaces the current allowlist. The service
keeps track of allowlist decisions per-profile.

Future CLs will also make the following changes:
- Allowlist entries will be saved to content settings when not in
incognito mode (same as desktop).
- Allowlist entries will have a timeout (same as desktop)
- Test methods like SetHttpPortForTesting() will be moved from the tab
helper to the service so that different tabs can read the same values
at once. This will allow us to test the feature with multiple tabs.

Bug: 1302509
Change-Id: Ia6e11991400de50a6136d579c7fc90ed8833a468
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3633920
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1001777}
NOKEYCHECK=True
GitOrigin-RevId: 791ebc7d84aa4f5d03a7f4c70813894d61834d69