[iOS] Fix crash in password manager settings
The solution consists simply in stating that, if the search mode is
on, “on device encryption” section should not be shown.

Note that it won’t appear when search end, which is not the expected
behavior.

This must also correct bug 1322464, as otherwise the DCHECK fail during the unit test.

Tested:
I added a unit test. I tested without the correction, and it crashes at the same place. I added the correction, and it succeed.

Manually Tested:
Open ipad with two windows, be logged-in, with password sync disabled and promotion of on-device experiment activated.
On one window, open password-manager and click on search.
On the second window, activate password sync and click done.

It should not crash anymore.

Fixed: 1322464
Bug: 1321057
Change-Id: I78a97358e06921848c9d9b6a076368a7c0ec24a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3627553
Commit-Queue: Arthur Milchior <arthurmilchior@google.com>
Auto-Submit: Arthur Milchior <arthurmilchior@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002118}
NOKEYCHECK=True
GitOrigin-RevId: ad57f289a804e215bfd4aae5ff52647f0af69bec