Update optimize_js.gni to support dependencies
Imported scripts will be resolved against the files listed in the
manifest files produced by the targets in `deps`. If the import is
not listed in any of the provided manifest files, it is assumed to exist
at `root_dir`.

Additionally, add support for an `output_name` which is passed to
rollup to specify the name of JS exports.

Bug: 1316376
Change-Id: I9aad9591d37e0e2541dde0a67f6bdedaf2c86b5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3627128
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1002127}
NOKEYCHECK=True
GitOrigin-RevId: eacaac5095b75a408bee5f9d0fabec0d409f0a65