[iOS] Use simulated request to load offline pages
With the new iOS 15 API, the offline page could be loaded directly
instead of loading the offline page URL, failing and then displaying
the offline version.

It is necessary to keep using `LoadData` method when there is a page
load with an associated offline version that fails. If using the
simulated request API, it means that the simulated request should use
the real URL of the page (so the error page URL) instead of the virtual
URL. So there is no real interest of only using the simulated request
in all cases.

Bug: 1323224
Change-Id: I5187733a5a895d11f9c64e4858c7cfd3427272a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645210
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1003083}
NOKEYCHECK=True
GitOrigin-RevId: e31af4e0ecc0fc48788607a8bedb92adac516657