Make sure screen object is created once in tests.
Replacing a screen instance is not safe because it carries a lot
of state. This CL moves all test screen initialization to test SetUp
phase so that single screen instance will be used during
a single test.

One exception is IOS because it currently relies on the implementation
that carries some state from one test to another, and I was unable to
eliminate that dependency. Looks like there is no IOS test that replaces
the screen instance, so it is not required, but someone who is familiar
with IOS needs to work on it if that's necessary.

A class ScopedNativeScreen is added: This class creates and sets a NativeScreen instance only when the screen is not already created (by tests)

Another class ScopedScreenOzone is added: This class creates and sets a ScreenOzone instance only when the screen is not already created (by tests). This is necessary as the creation of Screen may have to be delayed on some platform. I'll look into if we can fix or improve this separately.

2-pagers doc: https://docs.google.com/document/d/1q7lJnR06DhNtHYHZP8jpHzN19FUlcFWk7zDkyNQpxLU/edit?resourcekey=0-zO3n4kN3Ci1RLpvVAjpwAA

Bug: 1317416

Change-Id: I5cc878a855192ebb6143e3eae136c2f24525fa2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3247911
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Gary Kacmarcik <garykac@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Mike Wasserman <msw@chromium.org>
Commit-Queue: Mitsuru Oshima <oshima@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1004096}
NOKEYCHECK=True
GitOrigin-RevId: 94433a5877b9f4a49e06ef9923e8eda244077685