[ios] ARQuickLookTabHelper should not expose WebState accessor
Change ARQuickLookTabHelperDelegate to receive the WebState for
which a USDZ has been downloaded and needs to be presented. In
ARQuickLookCoordinator move the logic for presenting to an helper
class, use base::WeakPtr<web::WebState> to avoid Use-After-Free
and do not try to present to USDZ at the same time.

Bug: none
Change-Id: I6c1001017acba8be773e7bf71886747149f332b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3647059
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1004211}
NOKEYCHECK=True
GitOrigin-RevId: 39fc8e7f15e761dfa30afc43867c8e98406eeed0