[iOS] Make sure the callback is called a 2nd time
Prior to this CL, the boolean checking if the callback was called was
already YES before the frame was set. This hid the fact that the window
needed to be relaid out.

This CL adds the window relayout, and checks that the callback is called
after that relayout too.

Bug: 1318788
Change-Id: Iac44bbfd6b1e5121c993ac71eaed56efbc7fbc94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3650727
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Louis Romero <lpromero@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1004656}
NOKEYCHECK=True
GitOrigin-RevId: e90b899a740da97ca1c826dd576f12fc7b705b8d