[Shopping Service] Add tab helper to monitor web contents
This patch adds tab helpers for content and ios to watch for the
creation and destruction of web contents/state. These events and a
wrapped version of the web object are passed to the shopping service
so that it can interact with the web page.

Bug: 1322095
Change-Id: I63c4ecdad830895c5319a14e5648504eecce544f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3625902
Reviewed-by: David Trainor <dtrainor@chromium.org>
Reviewed-by: Zhiyuan Cai <zhiyuancai@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1004731}
NOKEYCHECK=True
GitOrigin-RevId: 35466e290c244a11d9700d4ab3a7191eb0fbb62d