[iOS] Interstitial page doesn't show Safe Browsing settings
The Safe Browsing interstitial page doesn't show the Safe Browsing
settings page when "Standard Protection" is enabled and the link
"turn on enhanced protection" is pressed. This only happens when a
WebState is inserted which we do not cover in the current
implementation. This CL adds logic to account for WebStates that are
inserted.

Fixed: 1322962
Change-Id: I60223e43b5c9cf83b478921db8adb6a410acaa0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3650839
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1004735}
NOKEYCHECK=True
GitOrigin-RevId: a8f2773db825e126c919a2ebc9121d08b260b20c