[Shopping Service] Move PDP metrics into service
This patch moves the PDP metrics into the shopping service which now
supports interactions with web contents. The new implementation is
functionally the same as the old but will work on all platforms if
enabled in finch.

The infra for testing has been separated to its own file so it can
be easily shared between suites. As such, tests for the metrics have
been added as well.

Bug: 1285460
Change-Id: Ie70fd4aa53195e4ca819a966bea68cd7ec34ed1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3639443
Reviewed-by: David Maunder <davidjm@chromium.org>
Reviewed-by: David Trainor <dtrainor@chromium.org>
Commit-Queue: Matthew Jones <mdjones@chromium.org>
Reviewed-by: Zhiyuan Cai <zhiyuancai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1004739}
NOKEYCHECK=True
GitOrigin-RevId: 71295ff3817cbb157d2b3cb93326ee5515866e7d