HTTPS-Only Mode: Clear the allowlist when browsing history is cleared
The HTTPS-Only Mode allowlist contains the list of sites that should not
be upgraded to HTTPS. A site is added to the allowlist if the user
clicks through the HTTPS-Only Mode interstitial.

The allowlist entries are saved to content settings with an expiry date
of 7 days after which they are ignored.

This CL clears the allowlist when browsing data is cleared from the
settings. The allowlist doesn't support clearing based on time ranges,
so any attempt to clear the browsing data will clear all allowlist
entries.

Bug: 1302509
Change-Id: Ib57dd446516dc1d519fab3241e324e4ed27146d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645606
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1005047}
NOKEYCHECK=True
GitOrigin-RevId: e381e5452f0f9c292068f0f3781c8274bb859371