[ios] Fix WebStateObserverTest.DownloadNavigation with new download API
The methods DidFinishNavigation()/DidStopLoading() were not called when
a navigation was transformed into a download when using the new API. It
was caught by WebStateObserverTest.DownloadNavigation which failed when
the flag was enabled. This may have been causing some issue with the UI
as not all WebStateObserver events were sent when a navigation turned
into a download on iOS 15.

Fix the regression by correctly calling the methods when the WKDownload
object is created by WKWebView and passed to CRWWKNavigationHandler.

Bug: 1252380
Change-Id: I123cf95d67349518f7f781bb5a38aa9fed232fed
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3650514
Reviewed-by: Joemer Ramos <joemerramos@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1005211}
NOKEYCHECK=True
GitOrigin-RevId: eabe124c547b313dfd7e8b44393fb9d6a2b53010