[ios] Fix sequence safety issue in ChromeActivityItemThumbnailGenerator
The ChromeActivityItemThumbnailGenerator instance may be deallocated on
a background thread managed by UIKit under certain circumstances. This
is unsafe because WebState is bound to the main UI thread sequence. It
thus crashed when the WebStateObserver was unregistered.

As the WebStateObserver was used to check for the destruction of the
WebState, it is simpler to use base::WeakPtr<web::WebState> which is
safe to destroy (but not access) on a background thread.

Add a unit test that dealloc a ChromeActivityItemThumbnailGenerator on
a background sequence, check that it failed with the old code and pass
with the new implementation.

Move the check for IsOffTheRecord() outside of the object initializer
as it allow to make the tests simpler (there is no need to create a
fake ChromeBrowserState).

Fixed: 1326659
Change-Id: I1210ced4ce7a6e4a281ae095b1a33b84573febeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3655346
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1005716}
NOKEYCHECK=True
GitOrigin-RevId: 5938e3414233044ac44257571be0f95228adeb72