[WebChannels] Fix crash when open overflow/popup menu in incognito
The DiscoverFeedService is started when luanch chrome, but it is only
for non-incognito browser. When calling from an incognito browser, it
can't get the correct instance as well as the metrics recorder.

This cl checks if the users are browsing in incognito before setting
the feed metrics recoder to the 3-dot menu.

Bug: 1327006
Change-Id: I43c60123de5c4766ff78a3530be6981ea599245c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645584
Reviewed-by: Adam Arcaro <adamta@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1005901}
NOKEYCHECK=True
GitOrigin-RevId: 29e66bd5f79cb1d7c98d7f6f3846883d137aeb1e