Eliminate base::LowerCaseEqualsASCII in /ios/chrome/browser
This CL is a behavior-preserving refactor. It replaces calls to
base::LowerCaseEqualsASCII with base::EqualsCaseInsensitiveASCII.

We are deprecating base::LowerCaseEqualsASCII because its interface is a
mild footgun, in that it requires the second parameter to be lowercase.

See the design doc for more details.
https://docs.google.com/document/d/18zWV0zN4oAqMZpond1Ut0QviHUeHCSwCo08mqThwHkQ/edit?usp=sharing

This CL was uploaded by git cl split.

R=sdefresne@chromium.org

Bug: 1326529
Change-Id: Iee2428caf50043db67adfadcf6b87855ac29d0a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3655945
Auto-Submit: Dan McArdle <dmcardle@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1006385}
NOKEYCHECK=True
GitOrigin-RevId: 6b766efe8c3e1754c2fe45d1e2d57975194a4327