HTTPS-Only Mode: Cleanup and fix interstitial reload
The current implementation of HTTPS-Only Mode tab helper on iOS has
many state variables which are set according to which phase of the
upgrade we are in. This is hard to keep track of, and buggy in some
cases. For example, the following scenario hits a DCHECK:
1. Open a site that hits the HTTPS-Only mode interstitial.
2. Allowlist the site
3. Clear the allowlist (this automatically reloads the tab and shows
the interstitial
4. Reload the interstitial.
The code assumes we can only get here if was_upgraded_ is true, which
is not the case.

This CL removes the boolean state variables and replaces them with an
enum that stores the current state. It also fixes the scenario above and
adds tests for various interstitial reload cases. The end result is
more readable and robust code.

Bug: 1302509
Change-Id: I8af6f8ea1d341f459b8c26462d7c678e659463e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3656040
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1006519}
NOKEYCHECK=True
GitOrigin-RevId: 269a3da30349d8a91a1229ad068ed34b2e3ed450