[iOS] Present the Autofill IPH
Depending on the feature engagement config, an IPH bubble is presented
when Autofill displays suggestions.

To anchor the bubble, this CL pipes the current layout guide center to
the Autofill view atop the keyboard, and tracks the first suggestion pill.

This CL adds a tap recognizer to the base view controller to catch taps outside
the bubble (dismissing the bubble).

The feature is behind the flag BubbleRichIPH with different states:
  - Disabled: no bubble, no suggestion highlight
  - Enabled (default bubble style)
  - Enabled Target Highlight (default bubble style)
  - Enabled Explicit dismissal
  - Enabled Dismissal and rich content
  - Enabled Dismissal, rich content, and snooze
Note: All the enabled states will enable the suggestion highlight, the
difference is in the style of the bubble.

This CL is based off of christianxu's crrev.com/c/3579325. Thank you Christian!

Specs:
https://www.figma.com/file/RNJcGlmnDcnpGmAhzfNnKy/Bling---IPH?node-id=447%3A84175

Bug: 1293370
Bug: 1310735
Change-Id: I43566bf255f853197dfbfde2d48851800d7534df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3645111
Commit-Queue: Louis Romero <lpromero@chromium.org>
Auto-Submit: Louis Romero <lpromero@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1006800}
NOKEYCHECK=True
GitOrigin-RevId: 2df26384021d09ba624e6ad9cd5fb8381cee0cb0