[WebChannels] Fix FollowJavaScriptFeature::HandleResponse crash
It crashes because -[NSArray addObject:] can't take a nil, and rarely
sometime the -[NSURL URLWithString:] returns nil because there are
non-ASCII characters in the string. So we need to check if url is a
valid NSURL object before we add it to the NSArray.

Bug: 1328017
Change-Id: I073e6b264d4278c480b5e1270bf64b42ff90be7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3665517
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1007293}
NOKEYCHECK=True
GitOrigin-RevId: 17ae670382d8ad3c4203f9a28f036f498767312c