chrome://policy: Create Web Components for policy cr.ui classes
Create Web Components for the various cr.ui.define() classes in
policy_base.js and serve them from the data source, but do not import
them into the UI yet (so they will not be used at runtime).

This is in preparation for replacing the existing cr.ui.define()
classes with these components in a followup CL.

This is a partial re-land of
https://chromium-review.googlesource.com/c/chromium/src/+/3648387

That CL was reverted due to a CrOS tast test issue. Since this CL
just adds the components and does not alter the structure of the page,
it will not have the same test failure. This CL is being broken off
to reduce the size of the change to keep in sync and number of
reviewers for the re-land.

Bug: 1322977
Change-Id: Iceb4b4225b65531d80b89b150fe803c108ea6996
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3671810
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Owen Min <zmin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1008413}
NOKEYCHECK=True
GitOrigin-RevId: f6b2f26a1a89c51b4b8938aa04e55a32b516ee04