HTTPS-Only Mode: Handle learn more link on iOS
Clicking on the "Learn More" link in the HTTPS-Only Mode on iOS
currently does nothing. This CL makes it open the help center article
in a new tab. This behavior is identical to the desktop and Android
versions of the interstitial.

Bug: 1302509
Change-Id: I4625209a5b56d3c5bbdacbdaf295773b805d74cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3671998
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1009298}
NOKEYCHECK=True
GitOrigin-RevId: 13a2edda2daca047210d0ce115ff9ae9c479f264