Update SigninPromoView to support NTP promo
Added support to compact promo style. (the image is on the leading side
of the promo).
Added support to non profile - non chromium icon -  image for cold mode.
Added support to text title on top of the text body of the promo.

Bug: 1328877
Change-Id: Ibb1cc4dd8d7a21447c1f27c392a109ab8457de33
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3665296
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1009680}
NOKEYCHECK=True
GitOrigin-RevId: e5217f04792512e082b2386d04c8ed6ed3625413