[iOS] [WebChannels] Fix feed header
Adds a z-position to the feed header to ensure that content suggestions
never overlap it. This was only visible with the content-suggestions-
header-migration experiment enabled.

Resets the sticky omnibox if it should not stick but was left in a
stuck state, like after rotating the NTP.

Bug: 1330594, 1329853, 1329828
Change-Id: I2a8fc1c17253299a918863a4580a0578d20401f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3684549
Reviewed-by: Tina Wang <tinazwang@chromium.org>
Commit-Queue: Adam Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/main@{#1010156}
NOKEYCHECK=True
GitOrigin-RevId: ee180baf86ad42413d077beb3175b900b2a71cfa