Implement Smart Sorting frecency algorithm.
Tracks destination usage from the new overflow menu and implements a
frecency algorithm to sort destinations shown in the overflow menu
carousel. The algorithm, at a high-level, works as follows:

(1) Divide the destinations carousel into two groups: (A) visible
"above-the-fold" destinations and (B) non-visible "below-the-fold"
destinations; "below-the-fold" destinations are made visible to the user
when they scroll the carousel.

(2) Get each destination's numClicks.

(3) Compare destination with highest numClicks in [group B] to
destination with lowest numClicks in [group A]

(4) Swap (i.e. "promote") the [group B] destination with the [group A]
one if B's numClicks exceeds A's.

Bug: 1322519
Change-Id: I530bb5f8d7b1679a0cb32272a77f6cbaa270db84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3663406
Commit-Queue: Benjamin Williams <bwwilliams@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/main@{#1010274}
NOKEYCHECK=True
GitOrigin-RevId: e3d5a487f4574145e039a4f755472c664acb869b