[iOS] Add settings UI for 3P intents in Incognito interstitial
This CL implements the UI for the new "Ask to Open Links from Other Apps
in Incognito" setting in the Privacy settings menu.

- String is added for the setting's text, new section and
switch item are conditionally added to the table view model if the flag
is enabled. Callback for info popover is added although it is not called
yet.
- Disabled versions of Incognito interstitial item and of
Incognito reauth item are added when Enterprise policy or device support
make the settings irrelevant. Incognito reauth item is now always shown
so as to be consistent with the behavior for the interstitial item.
- Incognito reauth footer item is replaced with an
informational popover. A new class is created for this purpose, which is
now responsible for implementing the "isPresentingFromButton" behavior,
and from which `EnterpriseInfoPopoverViewController` now inherits.
- Rewrote PrivacyTableViewController unit tests to
account for all permutations of Enhanced protection and 3P intents in
Incognito flag configurations.

Bug: 1318750
Change-Id: I7ac2ff5f7ec6f7cf370bbd0aec14c6b7106aa51a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3677004
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Quentin Pubert <qpubert@google.com>
Cr-Commit-Position: refs/heads/main@{#1010728}
NOKEYCHECK=True
GitOrigin-RevId: 0757c8bad4fbf0173ea9793a0e8f124b70133572